+++ src/main/kotlin/app/revanced/patches/youtube/misc/video/speed/remember/fingerprint/OnPlaybackSpeedItemClickFingerprint.kt
@@ -3,7 +3,7 @@ package app.revanced.patches.youtube.misc.video.speed.remember.fingerprint
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import org.jf.dexlib2.Opcode
 
-object OnPlaybackRateItemClickFingerprint : MethodFingerprint(
+object OnPlaybackSpeedItemClickFingerprint : MethodFingerprint(
     customFingerprint = { it.name == "onItemClick" },
     opcodes = listOf(
         Opcode.IGET_OBJECT,